+++ src/exactpy/models/financial/exchange_rate.py
@@ -1,26 +1,43 @@
-from sparkdantic import SparkField
-
 from exactpy.models.base import ExactOnlineBaseModel
 from exactpy.types import GUID, ExactUnixTimestamp
+from exactpy.utils.fields import GenericField
 
 
 class ExchangeRateModel(ExactOnlineBaseModel):
     _pk = "id"
-    id: GUID = SparkField(spark_type="string")
-    created: ExactUnixTimestamp = SparkField(spark_type="timestamp")
-    creator: GUID = SparkField(spark_type="string")
-    creator_full_name: str | None = SparkField(spark_type="string", default=None)
-    division: int | None = SparkField(spark_type="integer", default=None)
-    modified: ExactUnixTimestamp = SparkField(spark_type="timestamp")
-    modifier: GUID = SparkField(spark_type="string")
-    modifier_full_name: str | None = SparkField(spark_type="string", default=None)
-    rate: float | None = SparkField(spark_type="double", default=None)
-    source_currency: str | None = SparkField(spark_type="string", default=None)
-    source_currency_description: str | None = SparkField(
-        spark_type="string", default=None
+    id: GUID = GenericField(iceberg_type="string", spark_type="string")
+    created: ExactUnixTimestamp = GenericField(
+        iceberg_type="timestamp", spark_type="timestamp"
     )
-    start_date: ExactUnixTimestamp = SparkField(spark_type="timestamp")
-    target_currency: str | None = SparkField(spark_type="string", default=None)
-    target_currency_description: str | None = SparkField(
-        spark_type="string", default=None
+    creator: GUID = GenericField(iceberg_type="string", spark_type="string")
+    creator_full_name: str | None = GenericField(
+        iceberg_type="string", spark_type="string", default=None
+    )
+    division: int | None = GenericField(
+        iceberg_type="integer", spark_type="integer", default=None
+    )
+    modified: ExactUnixTimestamp = GenericField(
+        iceberg_type="timestamp", spark_type="timestamp"
+    )
+    modifier: GUID = GenericField(iceberg_type="string", spark_type="string")
+    modifier_full_name: str | None = GenericField(
+        iceberg_type="string", spark_type="string", default=None
+    )
+    rate: float | None = GenericField(
+        iceberg_type="double", spark_type="double", default=None
+    )
+    source_currency: str | None = GenericField(
+        iceberg_type="string", spark_type="string", default=None
+    )
+    source_currency_description: str | None = GenericField(
+        iceberg_type="string", spark_type="string", default=None
+    )
+    start_date: ExactUnixTimestamp = GenericField(
+        iceberg_type="timestamp", spark_type="timestamp"
+    )
+    target_currency: str | None = GenericField(
+        iceberg_type="string", spark_type="string", default=None
+    )
+    target_currency_description: str | None = GenericField(
+        iceberg_type="string", spark_type="string", default=None
     )