+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Unreleased
+
+### Fix
+
+- **models**: added icedantic integration for allowing iceberg schema exports
+- **controllers/base.py**: removed ABC override
+- added payroll namespace and activeemployment model and controller
+
 ## 0.0.23 (2025-11-19)
 
 ### Fix